fastapi-app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
torch==2.4.1
transformers==4.44.2
pillow==10.4.0
python-multipart==0.0.9